move preserve_float_range and do a basic refactor
