+++ notebooks/py/MERFISH_Pipeline_-_U2O2_Cell_Culture_-_1_FOV.py
@@ -95,7 +95,7 @@ high_passed = ghp.run(primary_image, verbose=True, in_place=False)
 # EPY: END markdown
 
 # EPY: START code
-dpsf = Filter.DeconvolvePSF(num_iter=15, sigma=2, clip=True)
+dpsf = Filter.DeconvolvePSF(num_iter=15, sigma=2)
 deconvolved = dpsf.run(high_passed, verbose=True, in_place=False)
 # EPY: END code
