+++ starfish/image/_filter/mean_high_pass.py
@@ -8,9 +8,10 @@ from scipy.ndimage.filters import uniform_filter
 from starfish.imagestack.imagestack import ImageStack
 from starfish.types import Number
 from starfish.util import click
+from starfish.util.dtype import preserve_float_range
 from ._base import FilterAlgorithmBase
 from .util import (
-    determine_axes_to_group_by, preserve_float_range, validate_and_broadcast_kernel_size
+    determine_axes_to_group_by, validate_and_broadcast_kernel_size
 )
