+++ starfish/image/_filter/util.py
@@ -1,7 +1,6 @@
 from typing import Set, Tuple, Union
 
 import numpy as np
-import xarray as xr
 from skimage.morphology import binary_opening, disk
 
 from starfish.types import Axes, Number
@@ -109,41 +108,6 @@ def validate_and_broadcast_kernel_size(
             valid_sigma = (sigma,) * 2
 
     return valid_sigma
-
-
-def preserve_float_range(
-        array: Union[xr.DataArray, np.ndarray],
-        rescale: bool=False) -> Union[xr.DataArray, np.ndarray]:
-    """
-    Clips values below zero to zero. If values above one are detected, clips them
-    to 1 unless `rescale` is True, in which case the input is scaled by
-    the max value and the dynamic range is preserved.
-
-    Parameters
-    ----------
-    array : Union[xr.DataArray, np.ndarray]
-        Array whose values should be in the interval [0, 1] but may not be.
-    rescale: bool
-        If true, scale values by the max.
-
-    Returns
-    -------
-    array : Union[xr.DataArray, np.ndarray]
-        Array whose values are in the interval [0, 1].
-
-    """
-    array = array.copy()
-
-    data = np.asarray(array)
-
-    if np.any(data < 0):
-        data[data < 0] = 0
-    if np.any(data > 1):
-        if rescale:
-            data /= data.max()
-        else:
-            data[data > 1] = 1
-    return array.astype(np.float32)
 
 
 def determine_axes_to_group_by(is_volume: bool) -> Set[Axes]:
