+++ starfish/intensity_table/intensity_table.py
@@ -8,7 +8,7 @@ import regional
 import xarray as xr
 
 from starfish.expression_matrix.expression_matrix import ExpressionMatrix
-from starfish.image._filter.util import preserve_float_range
+from starfish.util.dtype import preserve_float_range
 from starfish.types import Axes, Features, LOG, SpotAttributes, STARFISH_EXTRAS_KEY
